requirements.txt
# Core Web Framework
fastapi==0.121.2
uvicorn==0.38.0
starlette==0.49.3

# Data Validation
pydantic==2.12.4
pydantic_core==2.41.5

# Configuration & Environment
python-dotenv==1.2.1

# HTTP Requests
requests==2.32.5

# Database Clients
neo4j==6.0.3
qdrant-client==1.16.0
chromadb==1.3.4

# SQL Database
# Note: sqlite3 is included in Python standard library (no installation needed)
# For Memgraph installation:
# Docker: docker run -p 7687:7687 -p 7444:7444 memgraph/memgraph-platform
# Or install locally: https://memgraph.com/docs/getting-started/install-memgraph

# Machine Learning & AI
sentence-transformers==5.1.2
torch==2.9.0
transformers==4.57.1
numpy==2.2.6
huggingface-hub==0.36.0
safetensors==0.6.2
tokenizers==0.22.1

# LangChain Integration
langchain-core==0.3.79
langchain-huggingface==1.0.1
langchain-nvidia-ai-endpoints==0.3.19

# Development & Testing Tools
pyngrok==7.4.1
